refactor: break down large files into smaller, focused modules
- Modularized themes.ts into separate files for animation, sizes, color schemes, and factory functions
- Split AccessibleThemedComponents.tsx into individual component files
- Split ThemedComponents.tsx into individual component files
- Refactored AnalyticsService.ts into smaller modules (types, providers, helpers, etc.)
- Broke down AnalyticsSettingsScreen.tsx into smaller components

This improves code maintainability, testability, and makes the codebase easier to navigate.
