+++ apps/product/src/screens/Settings/Analytics/components/ActionButtons.tsx
@@ -1,0 +1,47 @@
+import React from "react";
+import { Linking, View } from "react-native";
+import { useButtonTracking } from "../../../../hooks/useAnalytics";
+import { AccessibleThemedButton } from "../../../../theme/components/AccessibleThemedComponents";
+import { useTheme } from "../../../../theme/ThemeContext";
+
+export const ActionButtons: React.FC = () => {
+	const { theme } = useTheme();
+	const trackPrivacyPolicy = useButtonTracking("privacy_policy_link");
+	const trackDataDeletion = useButtonTracking("request_data_deletion");
+
+	const handlePrivacyPolicy = () => {
+		trackPrivacyPolicy();
+		Linking.openURL("https://braingame.dev/privacy");
+	};
+
+	const handleDataDeletion = () => {
+		trackDataDeletion();
+		Linking.openURL("https://braingame.dev/data-deletion");
+	};
+
+	return (
+		<View style={{ gap: theme.sizes.spacingMD }}>
+			<AccessibleThemedButton
+				variant="outline"
+				size="medium"
+				fullWidth
+				onPress={handlePrivacyPolicy}
+				accessibilityLabel="View Privacy Policy"
+				accessibilityHint="Opens privacy policy in your browser"
+			>
+				View Privacy Policy
+			</AccessibleThemedButton>
+
+			<AccessibleThemedButton
+				variant="ghost"
+				size="medium"
+				fullWidth
+				onPress={handleDataDeletion}
+				accessibilityLabel="Request Data Deletion"
+				accessibilityHint="Opens data deletion request form in your browser"
+			>
+				Request Data Deletion
+			</AccessibleThemedButton>
+		</View>
+	);
+};
