+++ apps/product/src/services/analytics/constants.ts
@@ -1,0 +1,4 @@
+export const ANALYTICS_STORAGE_KEYS = {
+	ENABLED: "@braingame/analytics_enabled",
+	LAST_LAUNCH: "@braingame/last_launch",
+} as const;
