+++ apps/product/src/services/analytics/index.ts
@@ -1,0 +1,17 @@
+// Main exports
+export { analytics } from "./AnalyticsService";
+
+// Type exports
+export type { EventName, EventProperties, UserProperties, SuperProperties } from "./types";
+
+// Helper function exports
+export {
+	trackEvent,
+	trackScreen,
+	identifyUser,
+	setUserProperty,
+	setUserProperties,
+	trackTiming,
+	resetAnalytics,
+	flushAnalytics,
+} from "./helpers";
