+++ apps/product/src/services/analytics/providers/index.ts
@@ -1,0 +1,2 @@
+export type { AnalyticsProvider } from "./AnalyticsProvider";
+export { MockAnalyticsProvider } from "./MockProvider";
