+++ apps/product/src/services/analytics/utils.ts
@@ -1,0 +1,3 @@
+export function generateSessionId(): string {
+	return `${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
+}
