+++ apps/product/src/theme/colorSchemes/dark/default.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const defaultDark: Theme["colors"] = {
+	primary: "#4da3ff",
+	primaryLight: "#80c1ff",
+	primaryDark: "#0073e6",
+	onPrimary: "#000000",
+
+	secondary: "#868e96",
+	secondaryLight: "#adb5bd",
+	secondaryDark: "#495057",
+	onSecondary: "#000000",
+
+	accent: "#ff9999",
+	accentLight: "#ffcccc",
+	accentDark: "#ff6666",
+	onAccent: "#000000",
+
+	background: "#121212",
+	surface: "#1e1e1e",
+	surfaceVariant: "#2d2d2d",
+
+	text: "#ffffff",
+	textSecondary: "#adb5bd",
+	textDisabled: "#6c757d",
+
+	border: "#343a40",
+	borderLight: "#495057",
+	divider: "#495057",
+
+	success: "#52c41a",
+	warning: "#faad14",
+	error: "#ff4d4f",
+	info: "#1890ff",
+
+	onBackground: "#ffffff",
+	onSurface: "#ffffff",
+	onError: "#000000",
+
+	shadow: "#000000",
+	shadowLight: "#00000033",
+};
