+++ apps/product/src/theme/colorSchemes/dark/forest.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const forestDark: Theme["colors"] = {
+	primary: "#81c784",
+	primaryLight: "#b2fab4",
+	primaryDark: "#519657",
+	onPrimary: "#000000",
+
+	secondary: "#a5d6a7",
+	secondaryLight: "#d7ffd9",
+	secondaryDark: "#75a478",
+	onSecondary: "#000000",
+
+	accent: "#aed581",
+	accentLight: "#e1ffb1",
+	accentDark: "#7da453",
+	onAccent: "#000000",
+
+	background: "#0d1f0f",
+	surface: "#1b3a1d",
+	surfaceVariant: "#2e5431",
+
+	text: "#ffffff",
+	textSecondary: "#c8e6c9",
+	textDisabled: "#689f38",
+
+	border: "#2e5431",
+	borderLight: "#43734a",
+	divider: "#43734a",
+
+	success: "#8bc34a",
+	warning: "#ff9800",
+	error: "#e91e63",
+	info: "#03a9f4",
+
+	onBackground: "#ffffff",
+	onSurface: "#ffffff",
+	onError: "#000000",
+
+	shadow: "#000000",
+	shadowLight: "#00000033",
+};
