+++ apps/product/src/theme/colorSchemes/dark/midnight.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const midnightDark: Theme["colors"] = {
+	primary: "#7986cb",
+	primaryLight: "#aab6fe",
+	primaryDark: "#49599a",
+	onPrimary: "#000000",
+
+	secondary: "#b39ddb",
+	secondaryLight: "#e6ceff",
+	secondaryDark: "#836fa9",
+	onSecondary: "#000000",
+
+	accent: "#9fa8da",
+	accentLight: "#d1d9ff",
+	accentDark: "#6f79a8",
+	onAccent: "#000000",
+
+	background: "#0f0f1e",
+	surface: "#1a1a2e",
+	surfaceVariant: "#252542",
+
+	text: "#ffffff",
+	textSecondary: "#c5cae9",
+	textDisabled: "#5c6bc0",
+
+	border: "#252542",
+	borderLight: "#3f3f5c",
+	divider: "#3f3f5c",
+
+	success: "#81c784",
+	warning: "#ffb74d",
+	error: "#e57373",
+	info: "#64b5f6",
+
+	onBackground: "#ffffff",
+	onSurface: "#ffffff",
+	onError: "#000000",
+
+	shadow: "#000000",
+	shadowLight: "#00000033",
+};
