+++ apps/product/src/theme/colorSchemes/dark/ocean.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const oceanDark: Theme["colors"] = {
+	primary: "#4dd0e1",
+	primaryLight: "#88ffff",
+	primaryDark: "#009faf",
+	onPrimary: "#000000",
+
+	secondary: "#4db6ac",
+	secondaryLight: "#82e9de",
+	secondaryDark: "#00867d",
+	onSecondary: "#000000",
+
+	accent: "#80deea",
+	accentLight: "#b4ffff",
+	accentDark: "#4bacb8",
+	onAccent: "#000000",
+
+	background: "#0a1929",
+	surface: "#001e3c",
+	surfaceVariant: "#0a2744",
+
+	text: "#ffffff",
+	textSecondary: "#90caf9",
+	textDisabled: "#546e7a",
+
+	border: "#132f4c",
+	borderLight: "#1e4976",
+	divider: "#1e4976",
+
+	success: "#66bb6a",
+	warning: "#ffa726",
+	error: "#f44336",
+	info: "#29b6f6",
+
+	onBackground: "#ffffff",
+	onSurface: "#ffffff",
+	onError: "#000000",
+
+	shadow: "#000000",
+	shadowLight: "#00000033",
+};
