+++ apps/product/src/theme/colorSchemes/dark/sunset.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const sunsetDark: Theme["colors"] = {
+	primary: "#ffab40",
+	primaryLight: "#ffdd71",
+	primaryDark: "#c77c02",
+	onPrimary: "#000000",
+
+	secondary: "#ffcc80",
+	secondaryLight: "#ffffb0",
+	secondaryDark: "#ca9b52",
+	onSecondary: "#000000",
+
+	accent: "#ff7043",
+	accentLight: "#ffa270",
+	accentDark: "#c63f17",
+	onAccent: "#000000",
+
+	background: "#1a0f05",
+	surface: "#3e2723",
+	surfaceVariant: "#5d4037",
+
+	text: "#ffffff",
+	textSecondary: "#ffccbc",
+	textDisabled: "#a1887f",
+
+	border: "#5d4037",
+	borderLight: "#795548",
+	divider: "#795548",
+
+	success: "#7cb342",
+	warning: "#ffb300",
+	error: "#e53935",
+	info: "#039be5",
+
+	onBackground: "#ffffff",
+	onSurface: "#ffffff",
+	onError: "#000000",
+
+	shadow: "#000000",
+	shadowLight: "#00000033",
+};
