+++ apps/product/src/theme/colorSchemes/light/default.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const defaultLight: Theme["colors"] = {
+	primary: "#007fff",
+	primaryLight: "#4da3ff",
+	primaryDark: "#0059b3",
+	onPrimary: "#ffffff",
+
+	secondary: "#6c757d",
+	secondaryLight: "#868e96",
+	secondaryDark: "#495057",
+	onSecondary: "#ffffff",
+
+	accent: "#ff6b6b",
+	accentLight: "#ff9999",
+	accentDark: "#cc5555",
+	onAccent: "#ffffff",
+
+	background: "#ffffff",
+	surface: "#f8f9fa",
+	surfaceVariant: "#e9ecef",
+
+	text: "#212529",
+	textSecondary: "#6c757d",
+	textDisabled: "#adb5bd",
+
+	border: "#dee2e6",
+	borderLight: "#e9ecef",
+	divider: "#e9ecef",
+
+	success: "#28a745",
+	warning: "#ffc107",
+	error: "#dc3545",
+	info: "#17a2b8",
+
+	onBackground: "#212529",
+	onSurface: "#212529",
+	onError: "#ffffff",
+
+	shadow: "#000000",
+	shadowLight: "#00000008",
+};
