+++ apps/product/src/theme/colorSchemes/light/forest.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const forestLight: Theme["colors"] = {
+	primary: "#2e7d32",
+	primaryLight: "#60ad5e",
+	primaryDark: "#005005",
+	onPrimary: "#ffffff",
+
+	secondary: "#558b2f",
+	secondaryLight: "#85bb5c",
+	secondaryDark: "#255d00",
+	onSecondary: "#ffffff",
+
+	accent: "#8bc34a",
+	accentLight: "#bef67a",
+	accentDark: "#5a9216",
+	onAccent: "#000000",
+
+	background: "#f1f8e9",
+	surface: "#ffffff",
+	surfaceVariant: "#e8f5e9",
+
+	text: "#1b5e20",
+	textSecondary: "#33691e",
+	textDisabled: "#81c784",
+
+	border: "#c8e6c9",
+	borderLight: "#dcedc8",
+	divider: "#a5d6a7",
+
+	success: "#1b5e20",
+	warning: "#f57c00",
+	error: "#c62828",
+	info: "#01579b",
+
+	onBackground: "#1b5e20",
+	onSurface: "#1b5e20",
+	onError: "#ffffff",
+
+	shadow: "#000000",
+	shadowLight: "#00000008",
+};
