+++ apps/product/src/theme/colorSchemes/light/midnight.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const midnightLight: Theme["colors"] = {
+	primary: "#3f51b5",
+	primaryLight: "#757de8",
+	primaryDark: "#002984",
+	onPrimary: "#ffffff",
+
+	secondary: "#7c4dff",
+	secondaryLight: "#b47cff",
+	secondaryDark: "#3f1dcb",
+	onSecondary: "#ffffff",
+
+	accent: "#536dfe",
+	accentLight: "#8f9bff",
+	accentDark: "#0043ca",
+	onAccent: "#ffffff",
+
+	background: "#e8eaf6",
+	surface: "#ffffff",
+	surfaceVariant: "#c5cae9",
+
+	text: "#1a237e",
+	textSecondary: "#283593",
+	textDisabled: "#9fa8da",
+
+	border: "#9fa8da",
+	borderLight: "#c5cae9",
+	divider: "#7986cb",
+
+	success: "#4caf50",
+	warning: "#ff9800",
+	error: "#f44336",
+	info: "#2196f3",
+
+	onBackground: "#1a237e",
+	onSurface: "#1a237e",
+	onError: "#ffffff",
+
+	shadow: "#000000",
+	shadowLight: "#00000008",
+};
