+++ apps/product/src/theme/colorSchemes/light/ocean.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const oceanLight: Theme["colors"] = {
+	primary: "#006994",
+	primaryLight: "#4d95b8",
+	primaryDark: "#004d6d",
+	onPrimary: "#ffffff",
+
+	secondary: "#00838f",
+	secondaryLight: "#4db6ac",
+	secondaryDark: "#005662",
+	onSecondary: "#ffffff",
+
+	accent: "#00acc1",
+	accentLight: "#4dd0e1",
+	accentDark: "#007c91",
+	onAccent: "#ffffff",
+
+	background: "#f0f4f7",
+	surface: "#ffffff",
+	surfaceVariant: "#e1f5fe",
+
+	text: "#263238",
+	textSecondary: "#546e7a",
+	textDisabled: "#90a4ae",
+
+	border: "#cfd8dc",
+	borderLight: "#eceff1",
+	divider: "#b0bec5",
+
+	success: "#00c853",
+	warning: "#ffab00",
+	error: "#d50000",
+	info: "#0091ea",
+
+	onBackground: "#263238",
+	onSurface: "#263238",
+	onError: "#ffffff",
+
+	shadow: "#000000",
+	shadowLight: "#00000008",
+};
