+++ apps/product/src/theme/colorSchemes/light/sunset.ts
@@ -1,0 +1,42 @@
+import type { Theme } from "../../types";
+
+export const sunsetLight: Theme["colors"] = {
+	primary: "#ff6f00",
+	primaryLight: "#ff9e40",
+	primaryDark: "#c43e00",
+	onPrimary: "#000000",
+
+	secondary: "#f57c00",
+	secondaryLight: "#ffad42",
+	secondaryDark: "#bb4d00",
+	onSecondary: "#000000",
+
+	accent: "#ff5722",
+	accentLight: "#ff8a50",
+	accentDark: "#c41c00",
+	onAccent: "#ffffff",
+
+	background: "#fff3e0",
+	surface: "#ffffff",
+	surfaceVariant: "#ffe0b2",
+
+	text: "#e65100",
+	textSecondary: "#bf360c",
+	textDisabled: "#ffab91",
+
+	border: "#ffccbc",
+	borderLight: "#fbe9e7",
+	divider: "#ffab91",
+
+	success: "#558b2f",
+	warning: "#e65100",
+	error: "#b71c1c",
+	info: "#0277bd",
+
+	onBackground: "#e65100",
+	onSurface: "#e65100",
+	onError: "#ffffff",
+
+	shadow: "#000000",
+	shadowLight: "#00000008",
+};
