+++ apps/product/src/theme/components/accessible/index.ts
@@ -1,0 +1,5 @@
+export { AccessibleThemedView } from "./AccessibleThemedView";
+export { AccessibleThemedText } from "./AccessibleThemedText";
+export { AccessibleThemedButton } from "./AccessibleThemedButton";
+export { AccessibleThemedInput } from "./AccessibleThemedInput";
+export { AccessibleThemedSwitch } from "./AccessibleThemedSwitch";
