+++ apps/product/src/theme/components/accessible/utils.ts
@@ -1,0 +1,3 @@
+import type { TextStyle, ViewStyle } from "react-native";
+
+export type StyleProp = ViewStyle | TextStyle | (ViewStyle | TextStyle)[];
