+++ apps/product/src/theme/components/base/index.ts
@@ -1,0 +1,5 @@
+export { ThemedView } from "./ThemedView";
+export { ThemedText } from "./ThemedText";
+export { ThemedButton } from "./ThemedButton";
+export { ThemedInput } from "./ThemedInput";
+export { ThemedCard } from "./ThemedCard";
