+++ apps/product/src/theme/constants/animation.ts
@@ -1,0 +1,9 @@
+export const baseAnimation = {
+	durationFast: 150,
+	durationNormal: 300,
+	durationSlow: 500,
+	easeIn: "cubic-bezier(0.4, 0, 1, 1)",
+	easeOut: "cubic-bezier(0, 0, 0.2, 1)",
+	easeInOut: "cubic-bezier(0.4, 0, 0.2, 1)",
+	easeBounce: "cubic-bezier(0.68, -0.55, 0.265, 1.55)",
+};
