+++ apps/product/src/theme/constants/index.ts
@@ -1,0 +1,2 @@
+export { baseAnimation } from "./animation";
+export { baseSizes } from "./sizes";
