+++ apps/product/src/theme/constants/sizes.ts
@@ -1,0 +1,34 @@
+export const baseSizes = {
+	// Font sizes
+	fontSizeXS: 10,
+	fontSizeSM: 12,
+	fontSizeMD: 14,
+	fontSizeLG: 16,
+	fontSizeXL: 18,
+	fontSize2XL: 24,
+	fontSize3XL: 32,
+	fontSize4XL: 40,
+
+	// Spacing
+	spacingXS: 4,
+	spacingSM: 8,
+	spacingMD: 16,
+	spacingLG: 24,
+	spacingXL: 32,
+	spacing2XL: 48,
+	spacing3XL: 64,
+
+	// Border radius
+	radiusXS: 2,
+	radiusSM: 4,
+	radiusMD: 8,
+	radiusLG: 12,
+	radiusXL: 16,
+	radiusFull: 9999,
+
+	// Icon sizes
+	iconSM: 16,
+	iconMD: 24,
+	iconLG: 32,
+	iconXL: 48,
+};
