+++ apps/product/src/theme/factory/index.ts
@@ -1,0 +1,2 @@
+export { createTheme } from "./createTheme";
+export { createComponentStyles } from "./componentStyles";
